Performance improvement
I've factored out the recursive part of this function to a nested
function to avoid unecessarily repeating the initial (whole lists)
checks for every element in the list.

This was after feedback from people on Zulip.

There's also other ways to improve this code using pattern matching
and it seems that the whole function can be replaced with the equals
operator in Roc which I've been told will work on a list of strings.